+++ backend/s32x-core/src/audio.rs
@@ -2,7 +2,9 @@ mod constants;
 
 use crate::api::Sega32XEmulatorConfig;
 use bincode::{Decode, Encode};
+use genesis_core::GenesisLowPassFilter;
 use genesis_core::audio::{GenesisAudioFilter, Ym2612Resampler};
+use jgenesis_common::audio::iir::FirstOrderIirFilter;
 use jgenesis_common::audio::{CubicResampler, DEFAULT_OUTPUT_FREQUENCY, FirResampler};
 use jgenesis_common::frontend::{AudioOutput, TimingMode};
 use smsgg_core::audio::PsgResampler;
@@ -12,15 +14,103 @@ const PAL_GENESIS_MCLK_FREQUENCY: f64 = genesis_core::audio::PAL_GENESIS_MCLK_FR
 
 const PSG_COEFFICIENT: f64 = genesis_core::audio::PSG_COEFFICIENT;
 
+// -2 dB (10^(-2 / 20))
+const PWM_COEFFICIENT: f64 = 0.7943282347242815;
+
+fn new_pwm_48khz_low_pass() -> FirstOrderIirFilter {
+    // First-order Butterworth IIR filter targeting 3390 Hz cutoff with source frequency of 48000 Hz
+    FirstOrderIirFilter::new(&[0.18406577751250913, 0.18406577751250913], &[
+        1.0,
+        -0.6318684449749816,
+    ])
+}
+
+fn new_pwm_44khz_low_pass() -> FirstOrderIirFilter {
+    // First-order Butterworth IIR filter targeting 3390 Hz cutoff with source frequency of 44100 Hz
+    FirstOrderIirFilter::new(&[0.1976272152714313, 0.1976272152714313], &[1.0, -0.6047455694571374])
+}
+
+// This silliness is necessary to handle dynamic resampling ratio; the frontend doesn't indicate
+// whether the output frequency is the original frequency or dynamic-adjusted
+fn round_output_frequency(output_frequency: u64) -> u64 {
+    let diff_48khz = (output_frequency as i64 - 48000).abs();
+    let diff_44khz = (output_frequency as i64 - 44100).abs();
+
+    if diff_48khz <= diff_44khz { 48000 } else { 44100 }
+}
+
+fn new_pwm_low_pass(output_frequency: u64) -> FirstOrderIirFilter {
+    match output_frequency {
+        48000 => new_pwm_48khz_low_pass(),
+        44100 => new_pwm_44khz_low_pass(),
+        _ => panic!(
+            "new_pwm_low_pass(freq) should only be called with 48000 or 44100, was {output_frequency}"
+        ),
+    }
+}
+
+#[derive(Debug, Clone, Encode, Decode)]
+struct PwmAudioFilter {
+    gen_low_pass_setting: GenesisLowPassFilter,
+    apply_gen_lpf_to_pwm: bool,
+    rounded_output_frequency: u64,
+    pwm_lpf_l: FirstOrderIirFilter,
+    pwm_lpf_r: FirstOrderIirFilter,
+}
+
+impl PwmAudioFilter {
+    fn new(config: &Sega32XEmulatorConfig, output_frequency: u64) -> Self {
+        let rounded_output_frequency = round_output_frequency(output_frequency);
+        Self {
+            gen_low_pass_setting: config.genesis.low_pass,
+            apply_gen_lpf_to_pwm: config.apply_genesis_lpf_to_pwm,
+            rounded_output_frequency,
+            pwm_lpf_l: new_pwm_low_pass(rounded_output_frequency),
+            pwm_lpf_r: new_pwm_low_pass(rounded_output_frequency),
+        }
+    }
+
+    fn filter(&mut self, (sample_l, sample_r): (f64, f64)) -> (f64, f64) {
+        if !self.apply_gen_lpf_to_pwm {
+            return (sample_l, sample_r);
+        }
+
+        (self.pwm_lpf_l.filter(sample_l), self.pwm_lpf_r.filter(sample_r))
+    }
+
+    fn reload_config(&mut self, config: &Sega32XEmulatorConfig) {
+        if self.gen_low_pass_setting == config.genesis.low_pass
+            && self.apply_gen_lpf_to_pwm == config.apply_genesis_lpf_to_pwm
+        {
+            return;
+        }
+
+        *self = Self::new(config, self.rounded_output_frequency);
+    }
+
+    fn update_output_frequency(&mut self, output_frequency: u64) {
+        let rounded_output_frequency = round_output_frequency(output_frequency);
+        if self.rounded_output_frequency == rounded_output_frequency {
+            return;
+        }
+
+        self.rounded_output_frequency = rounded_output_frequency;
+        self.pwm_lpf_l = new_pwm_low_pass(output_frequency);
+        self.pwm_lpf_r = new_pwm_low_pass(output_frequency);
+    }
+}
+
 #[derive(Debug, Clone, Encode, Decode)]
 pub struct PwmResampler {
+    filter: PwmAudioFilter,
     resampler: CubicResampler,
     lpf: FirResampler<{ constants::PWM_LPF_TAPS }, 0>,
 }
 
 impl PwmResampler {
-    pub fn new() -> Self {
+    pub fn new(config: &Sega32XEmulatorConfig, output_frequency: u64) -> Self {
         Self {
+            filter: PwmAudioFilter::new(config, output_frequency),
             resampler: CubicResampler::new(22000.0),
             lpf: FirResampler::new(
                 DEFAULT_OUTPUT_FREQUENCY as f64,
@@ -33,6 +123,7 @@ impl PwmResampler {
     pub fn collect_sample(&mut self, sample_l: f64, sample_r: f64) {
         self.resampler.collect_sample(sample_l, sample_r);
         while let Some((output_l, output_r)) = self.resampler.output_buffer_pop_front() {
+            let (output_l, output_r) = self.filter.filter((output_l, output_r));
             self.lpf.collect_sample(output_l, output_r);
         }
     }
@@ -49,7 +140,12 @@ impl PwmResampler {
         self.resampler.update_source_frequency(source_frequency);
     }
 
+    fn reload_config(&mut self, config: &Sega32XEmulatorConfig) {
+        self.filter.reload_config(config);
+    }
+
     pub fn update_output_frequency(&mut self, output_frequency: u64) {
+        self.filter.update_output_frequency(output_frequency);
         self.resampler.update_output_frequency(output_frequency);
         self.lpf.update_source_frequency(output_frequency as f64);
         self.lpf.update_output_frequency(output_frequency);
@@ -78,7 +174,7 @@ impl Sega32XResampler {
             gen_filter: GenesisAudioFilter::new(config.genesis.low_pass),
             ym2612_resampler: genesis_core::audio::new_ym2612_resampler(genesis_mclk_frequency),
             psg_resampler: smsgg_core::audio::new_psg_resampler(genesis_mclk_frequency),
-            pwm_resampler: PwmResampler::new(),
+            pwm_resampler: PwmResampler::new(&config, 48000),
             ym2612_enabled: config.genesis.ym2612_enabled,
             psg_enabled: config.genesis.psg_enabled,
             pwm_enabled: config.pwm_enabled,
@@ -122,8 +218,10 @@ impl Sega32XResampler {
                 self.pwm_enabled,
             );
 
-            let sample_l = (ym2612_l + PSG_COEFFICIENT * psg_l + pwm_l).clamp(-1.0, 1.0);
-            let sample_r = (ym2612_r + PSG_COEFFICIENT * psg_r + pwm_r).clamp(-1.0, 1.0);
+            let sample_l =
+                (ym2612_l + PSG_COEFFICIENT * psg_l + PWM_COEFFICIENT * pwm_l).clamp(-1.0, 1.0);
+            let sample_r =
+                (ym2612_r + PSG_COEFFICIENT * psg_r + PWM_COEFFICIENT * pwm_r).clamp(-1.0, 1.0);
 
             audio_output.push_sample(sample_l, sample_r)?;
         }
@@ -137,6 +235,7 @@ impl Sega32XResampler {
         self.pwm_enabled = config.pwm_enabled;
 
         self.gen_filter.reload_config(&config.genesis);
+        self.pwm_resampler.reload_config(&config);
     }
 
     pub fn update_output_frequency(&mut self, output_frequency: u64) {